+++ server/src/main/java/org/opensearch/plugins/SearchPlugin.java
@@ -96,6 +96,7 @@ import static java.util.Collections.emptyMap;
  * @opensearch.api
  */
 public interface SearchPlugin {
+
     /**
      * The new {@link ScoreFunction}s defined by this plugin.
      */