+++ server/src/main/java/org/opensearch/search/internal/ContextIndexSearcher.java
@@ -152,8 +152,12 @@ public class ContextIndexSearcher extends IndexSearcher implements Releasable {
         this.searchContext = searchContext;
     }
 
-    public void setProfiler(QueryProfiler profiler) {
+    public void setQueryProfiler(QueryProfiler profiler) {
         this.profiler = profiler;
+    }
+
+    public QueryProfiler getQueryProfiler() {
+        return profiler;
     }
 
     /**
@@ -213,7 +217,12 @@ public class ContextIndexSearcher extends IndexSearcher implements Releasable {
             // createWeight() is called for each query in the tree, so we tell the queryProfiler
             // each invocation so that it can build an internal representation of the query
             // tree
-            ContextualProfileBreakdown<QueryTimingType> profile = profiler.getQueryBreakdown(query);
+            ContextualProfileBreakdown profile = null;
+            try {
+                profile = profiler.getQueryBreakdown(query);
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
             Timer timer = profile.getTimer(QueryTimingType.CREATE_WEIGHT);
             timer.start();
             final Weight weight;