+++ server/src/main/java/org/opensearch/search/profile/AbstractProfileBreakdown.java
@@ -32,9 +32,11 @@
 
 package org.opensearch.search.profile;
 
-import java.util.Collections;
+import org.opensearch.common.annotation.PublicApi;
+
 import java.util.HashMap;
 import java.util.Map;
+import java.util.TreeMap;
 
 import static java.util.Collections.emptyMap;
 
@@ -45,44 +47,55 @@ import static java.util.Collections.emptyMap;
  *
  * @opensearch.internal
  */
-public abstract class AbstractProfileBreakdown<T extends Enum<T>> {
+@PublicApi(since = "3.0.0")
+public abstract class AbstractProfileBreakdown {
 
-    /**
-     * The accumulated timings for this query node
-     */
-    protected final Timer[] timings;
-    protected final T[] timingTypes;
-    public static final String TIMING_TYPE_COUNT_SUFFIX = "_count";
-    public static final String TIMING_TYPE_START_TIME_SUFFIX = "_start_time";
+    public static final String NODE_TIME_RAW = "time_in_nanos";
+
+    private final Map<String, ProfileMetric> metrics;
 
     /** Sole constructor. */
-    public AbstractProfileBreakdown(Class<T> clazz) {
-        this.timingTypes = clazz.getEnumConstants();
-        timings = new Timer[timingTypes.length];
-        for (int i = 0; i < timings.length; ++i) {
-            timings[i] = new Timer();
+    public AbstractProfileBreakdown(Map<String, Class<? extends ProfileMetric>> metricClasses) {
+        Map<String, ProfileMetric> map = new HashMap<>();
+        for (Map.Entry<String, Class<? extends ProfileMetric>> entry : metricClasses.entrySet()) {
+            try {
+                map.put(entry.getKey(), entry.getValue().getConstructor(String.class).newInstance(entry.getKey()));
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
         }
+        this.metrics = map;
     }
 
-    public Timer getTimer(T timing) {
-        return timings[timing.ordinal()];
+    public Timer getTimer(Enum<?> type) {
+        ProfileMetric metric = metrics.get(type.toString());
+        assert metric instanceof Timer : "Metric " + type + " is not a timer";
+        return (Timer) metric;
     }
 
-    public void setTimer(T timing, Timer timer) {
-        timings[timing.ordinal()] = timer;
+    public ProfileMetric getMetric(String name) {
+        return metrics.get(name);
     }
 
     /**
-     * Build a timing count breakdown for current instance
+     * Build a breakdown for current instance
      */
     public Map<String, Long> toBreakdownMap() {
-        Map<String, Long> map = new HashMap<>(this.timings.length * 3);
-        for (T timingType : this.timingTypes) {
-            map.put(timingType.toString(), this.timings[timingType.ordinal()].getApproximateTiming());
-            map.put(timingType + TIMING_TYPE_COUNT_SUFFIX, this.timings[timingType.ordinal()].getCount());
-            map.put(timingType + TIMING_TYPE_START_TIME_SUFFIX, this.timings[timingType.ordinal()].getEarliestTimerStartTime());
+        Map<String, Long> map = new TreeMap<>();
+        for (Map.Entry<String, ProfileMetric> entry : metrics.entrySet()) {
+            map.putAll(entry.getValue().toBreakdownMap());
         }
-        return Collections.unmodifiableMap(map);
+        return map;
+    }
+
+    public long toNodeTime() {
+        long total = 0;
+        for (Map.Entry<String, ProfileMetric> entry : metrics.entrySet()) {
+            if (entry.getValue() instanceof Timer) {
+                total += ((Timer) entry.getValue()).getApproximateTiming();
+            }
+        }
+        return total;
     }
 
     /**
@@ -92,11 +105,4 @@ public abstract class AbstractProfileBreakdown<T extends Enum<T>> {
         return emptyMap();
     }
 
-    public long toNodeTime() {
-        long total = 0;
-        for (T timingType : timingTypes) {
-            total += timings[timingType.ordinal()].getApproximateTiming();
-        }
-        return total;
-    }
 }