+++ server/src/main/java/org/opensearch/search/profile/ContextualProfileBreakdown.java
@@ -10,27 +10,27 @@ package org.opensearch.search.profile;
 
 import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.search.Collector;
+import org.opensearch.common.annotation.PublicApi;
 
 import java.util.List;
 import java.util.Map;
 
 /**
- * Provide contextual profile breakdowns which are associated with freestyle context. Used when concurrent
- * search over segments is activated and each collector needs own non-shareable profile breakdown instance.
- *
- * @opensearch.internal
+ * A {@link AbstractProfileBreakdown} for query timings with contexts.
  */
-public abstract class ContextualProfileBreakdown<T extends Enum<T>> extends AbstractProfileBreakdown<T> {
-    public ContextualProfileBreakdown(Class<T> clazz) {
-        super(clazz);
-    }
+@PublicApi(since = "3.0.0")
+public abstract class ContextualProfileBreakdown extends AbstractProfileBreakdown {
 
     /**
-     * Return (or create) contextual profile breakdown instance
-     * @param context freestyle context
-     * @return contextual profile breakdown instance
+     * Sole constructor.
+     *
+     * @param metrics
      */
-    public abstract AbstractProfileBreakdown<T> context(Object context);
+    public ContextualProfileBreakdown(Map<String, Class<? extends ProfileMetric>> metrics) {
+        super(metrics);
+    }
+
+    public abstract ContextualProfileBreakdown context(Object context);
 
     public void associateCollectorToLeaves(Collector collector, LeafReaderContext leaf) {}