+++ server/src/main/java/org/opensearch/search/profile/query/InternalQueryProfileTree.java
@@ -32,6 +32,7 @@
 
 package org.opensearch.search.profile.query;
 
+import org.apache.lucene.search.Query;
 import org.opensearch.search.profile.ContextualProfileBreakdown;
 import org.opensearch.search.profile.ProfileResult;
 
@@ -43,7 +44,7 @@ import org.opensearch.search.profile.ProfileResult;
 public class InternalQueryProfileTree extends AbstractQueryProfileTree {
 
     @Override
-    protected ContextualProfileBreakdown<QueryTimingType> createProfileBreakdown() {
-        return new QueryProfileBreakdown();
+    protected ContextualProfileBreakdown createProfileBreakdown(Query query) {
+        return new QueryProfileBreakdown(QueryProfileBreakdown.getQueryTimers());
     }
 }