+++ server/src/main/java/org/opensearch/search/profile/query/QueryProfiler.java
@@ -53,7 +53,7 @@ import java.util.Objects;
  * @opensearch.api
  */
 @PublicApi(since = "1.0.0")
-public class QueryProfiler extends AbstractProfiler<ContextualProfileBreakdown<QueryTimingType>, Query> {
+public class QueryProfiler extends AbstractProfiler<ContextualProfileBreakdown, Query> {
 
     /**
      * The root Collector used in the search
@@ -102,6 +102,10 @@ public class QueryProfiler extends AbstractProfiler<ContextualProfileBreakdown<Q
      */
     public CollectorResult getCollector() {
         return collector.getCollectorTree();
+    }
+
+    public ContextualProfileBreakdown getTopBreakdown() {
+        return profileTree.getStackTop();
     }
 
 }